pyproject.toml
[tool.poetry]
name = "gang-du-pos"
version = "0.1.0"
description = ""
authors = ["JackXu19930319 <49975224+JackXu19930319@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.29.0"
fastapi = "^0.110.2"
gunicorn = "^22.0.0"
mysql-connector-python = "^8.3.0"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
passlib = "^1.7.4"
pillow = "^10.3.0"
requests = "^2.31.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
